Search Bar fully made
