+++ components/searchBar/searchBar.js
@@ -1,61 +1,81 @@
 import React, { useState } from 'react';
-import { View, TextInput, Button } from 'react-native';
+import { View,Text, TextInput, Button, TouchableOpacity, FlatList} from 'react-native';
 import { icon } from 'react-native';
+import  styles  from './searchBar.style.js';
+import { MaterialIcons } from '@expo/vector-icons';
 
-const SearchBarWithFilters = ({ onSearch }) => {
+
+
+const SearchBar = ({ onSearch }) => {
   const [searchText, setSearchText] = useState('');
-  const [selectedFilters, setSelectedFilters] = useState([]);
+  const [data, setData] = useState([
+    { id: 1, name: 'Onion' },
+    { id: 2, name: 'Tomato' },
+    { id: 3, name: 'Potato' },
+    { id: 4, name: 'Pepper' },
+    { id: 5, name: 'Beef' },
+    { id: 6, name: 'Chicken' },
+    { id: 7, name: 'Mince' },
+    { id: 8, name: 'Rice' },
+    { id: 9, name: 'Spaghetti' },
+    { id: 10, name: 'Macaroni' },
+  ]);
+  const [filteredData, setFilteredData] = useState([]);
 
   const handleSearch = () => {
-    // Call the onSearch callback with the search text and selected filters
-    onSearch(searchText, selectedFilters);
+    const filteredResults = data.filter(item =>
+      item.name.toLowerCase().includes(searchText.toLowerCase())
+    );
+    setFilteredData(filteredResults);
   };
 
-  const toggleFilter = (filter) => {
-    const updatedFilters = [...selectedFilters];
-    const index = updatedFilters.indexOf(filter);
-    if (index > -1) {
-      updatedFilters.splice(index, 1); // Remove filter if already selected
-    } else {
-      updatedFilters.push(filter); // Add filter if not selected
-    }
-    setSelectedFilters(updatedFilters);
+  const handleFilter = filter => {
+    const filteredResults = data.filter(item => item.name.toLowerCase() === filter.toLowerCase());
+    setFilteredData(filteredResults);
   };
+
+  const handleReset = () => {
+    setFilteredData([]);
+    setSearchText('');
+  };
+
+  const renderItem = ({ item }) => {
+    return <Text>{item.name}</Text>;
+  };
+
+  
 
   return (
-    <View>
-      <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+    <View style={styles.container}>
+      <View style={styles.searchContainer}>
         <TextInput
-          style={{ flex: 1, height: 40, borderColor: 'gray', borderWidth: 1, padding: 10 }}
+          style={styles.searchInput}
           placeholder="Search"
+          placeholderTextColor="#333"
           value={searchText}
-          onChangeText={(text) => setSearchText(text)}
+          onChangeText={text => setSearchText(text)}
         />
-        <Button title="Search" onPress={handleSearch} />
+        <TouchableOpacity onPress={handleSearch} style={styles.searchIconContainer}>
+          <MaterialIcons name="search" size={24} color="white" />
+        </TouchableOpacity>
+        <Button title="Reset" onPress={handleReset} />
       </View>
-      <View style={{ flexDirection: 'row', flexWrap: 'wrap', marginTop: 10 }}>
-        {['Tomato',
-          'Onion',
-          'Pepper', 
-          'Eggs', 
-          'Garlic', 
-          'Rice',
-          'Chicken',
-          'Beef',
-          'Pork', 
-          'Spaghetti',
-          'Macaroni',
-        ].map((filter) => (
-          <Icon
-            key={filter}
-            name={selectedFilters.includes(filter) ? 'check-box' : 'check-box-outline-blank'}
-            onPress={() => toggleFilter(filter)}
-            containerStyle={{ marginRight: 10 }}
+      <View style={styles.filterContainer}>
+        {data.map(item => (
+          <Button
+            key={item.id}
+            title={item.name}
+            onPress={() => handleFilter(item.name)}
+            style={styles.filterButton}
           />
         ))}
       </View>
+      <FlatList
+        data={filteredData.length > 0 ? filteredData : data}
+        renderItem={renderItem}
+        keyExtractor={item => item.id.toString()}
+      />
     </View>
   );
 };
-
-export default searchBar;
+export default SearchBar;
